+++ frontend/src/types/post/post.dto.ts
@@ -1,9 +1,18 @@
+
+
+interface PostActionDto {
+    id: number;
+    userId: number;
+    postId: number;
+}
+
+
 export interface PostDto {
     id: number;
     title: string;
     content: string;
     image: string;
-    likesCount: number;
-    isLiked: boolean;
     userId: number;
+    likes: PostActionDto[];
+    dislikes: PostActionDto[];
 }